requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==5.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
frozenlist==1.7.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.1
multidict==6.6.4
openai==1.3.7
packaging==25.0
Pillow==10.1.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
twilio==8.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
yarl==1.20.1
